requirements.txt
Flask==3.1.2
Flask-Session==0.5.0
Pillow==12.1.0
python-dotenv==1.2.1
bcrypt==4.1.2
requests==2.32.5
bleach==6.3.0
